% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj2char.R
\name{adj2char}
\alias{adj2char}
\title{Flatten a list of adjacency matrices into a character matrix}
\usage{
adj2char(matlist, adj_prefix = "IG", ...)
}
\arguments{
\item{matlist}{a list of adjacency matrices}

\item{adj_prefix}{prefix to add to the output row names}

\item{...}{arguments passed to the matrix flattening utility function}
}
\value{
a binary character matrix indicating the presence (1) or absence (0) of a link
between vertices
}
\description{
Flatten a list of adjacency matrices into a character matrix
}
\details{
Assumes that the adjacency matrix has head and tail vertices in the columns and rows,
respectively.
}
\examples{
set.seed(9823)
n <- 3  # set number of species
As <- replicate(5, -diag(n) - 1.5 * matrix(runif(n^2), n, n), simplify = FALSE)
r <- matrix(1, n, 1)
schs <- lapply(As, inv_scheme, r)
gras <- lapply(schs, inv_graph)
pgras <- pad_adj_mats(gras)
adj2char(pgras)
}
