% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_paths.R
\name{count_paths}
\alias{count_paths}
\title{Count invasion graph paths between nodes}
\usage{
count_paths(IG, from = "EMPTY", to)
}
\arguments{
\item{IG}{invasion graph, either raw or tidy}

\item{from}{name of the source vertex/node in the invasion graph}

\item{to}{name of the target vertex/node in the invasion graph}
}
\value{
an integer count of the total paths between nodes
}
\description{
Count invasion graph paths between nodes
}
\examples{
n <- 6  # set number of species
A <- -diag(n) - 1.5 * matrix(runif(n^2), n, n)
r <- matrix(1, n, 1)
sch <- inv_scheme(A, r)
gra <- inv_graph(IS = sch)
tidy_gra <- tidy_inv_graph(gra)
count_paths(tidy_gra, from = "EMPTY", to = "345")

}
\author{
William K. Petry
}
