% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_states.R
\name{end_states}
\alias{end_states}
\title{Find the end state or states of an invasion graph.}
\usage{
end_states(dat)
}
\arguments{
\item{dat}{list containing species composition(s) of end state(s) and their corresponding
-i subcommunities}
}
\value{
list containing the compositions of the end states of community assembly
}
\description{
Find the end state or states of an invasion graph.
}
\examples{
n <- 6  # set number of species
A <- -diag(n) - 1.5 * matrix(runif(n^2), n, n)
r <- matrix(1, n, 1)
sch <- inv_scheme(A, r)
gra <- inv_graph(IS = sch)
end_states(gra)

}
\author{
Sebastian Schreiber

William K. Petry
}
