% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_matrix.R
\name{flatten_matrix}
\alias{flatten_matrix}
\title{Flatten a matrix into a vector}
\usage{
flatten_matrix(x, by = c("col", "row"), name_sep = "->")
}
\arguments{
\item{x}{a matrix, typically with row and column names}

\item{by}{whether to flatten the matrix into a vector column-wise or row-wise}

\item{name_sep}{character(s) deliminator used between row and column names in flattened output}
}
\value{
a named vector
}
\description{
Flatten a matrix into a vector
}
\examples{
mat <- matrix(runif(9), nrow = 3, dimnames = list(LETTERS[1:3], letters[1:3]))
flatten_matrix(mat)
flatten_matrix(mat, by = "row", name_sep = "_")

}
