% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_graph.R
\name{inv_graph}
\alias{inv_graph}
\title{Construct the invasion graph corresponding to an invasion scheme}
\usage{
inv_graph(IS, skip_minus_i = FALSE)
}
\arguments{
\item{IS}{matrix containing an invasion scheme with rows representing communities and
columns representing species.}

\item{skip_minus_i}{logical, when TRUE the -i communities will be skipped, speeding up
the calculation}
}
\value{
list containing an invasion graph and corresponding metadata. See Details for
    description of list elements.
}
\description{
Construct the invasion graph corresponding to an invasion scheme
}
\details{
The returned list includes the following named elements:
\describe{
  \item{IG}{matrix containing the invasion graph}
  \item{acyclic}{logical, whether the invasion graph is acyclic}
  \item{permanent}{logical matrix indicating the permanence of each community}
  \item{minus.i}{list containing the indices of each community that is a -i community}
  \item{number.species}{integer vector of the number of species in each community}
  \item{composition}{list containing vectors with the species ids in each community}
  \item{IS}{matrix containing the originally supplied invasion scheme}
}
}
\examples{
n <- 6  # set number of species
A <- -diag(n) - 1.5 * matrix(runif(n^2), n, n)
r <- matrix(1, n, 1)
sch <- inv_scheme(A, r)
gra <- inv_graph(IS = sch)
gra2 <- inv_graph(IS = sch, skip_minus_i = TRUE)

}
