% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_adj_mats.R
\name{pad_adj_mats}
\alias{pad_adj_mats}
\title{Homogenize the dimensions of a set of invasion graph adjacency matrices}
\usage{
pad_adj_mats(iglist)
}
\arguments{
\item{iglist}{A list of objects returned by `inv_graph`}
}
\value{
a list of adjacency matrices with identical vertices
}
\description{
Homogenize the dimensions of a set of invasion graph adjacency matrices
}
\examples{
set.seed(9823)
n <- 3  # set number of species
As <- replicate(5, -diag(n) - 1.5 * matrix(runif(n^2), n, n), simplify = FALSE)
r <- matrix(1, n, 1)
schs <- lapply(As, inv_scheme, r)
gras <- lapply(schs, inv_graph)
(pgras <- pad_adj_mats(gras))

sapply(gras, function(x) nrow(x$IG))  # variable number of communities in original IGs
sapply(pgras, function(x) nrow(x))  # identical number of communities in padded IGs

}
