% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inv_graph.R
\name{tidy_inv_graph}
\alias{tidy_inv_graph}
\title{Tidy invasion graph}
\usage{
tidy_inv_graph(dat, omit = c(), bend = 0.75)
}
\arguments{
\item{dat}{list containing an invasion graph and corresponding metadata}

\item{omit}{character vector of communities to omit from the plot}

\item{bend}{numeric, degree to bend nodes of the same diversity (0 is linear, 1 is maximum)}
}
\value{
a tibble representation of the invasion graph, with added metadata
}
\description{
Tidy invasion graph
}
\examples{
n <- 6  # set number of species
A <- -diag(n) - 1.5 * matrix(runif(n^2), n, n)
r <- matrix(1, n, 1)
sch <- inv_scheme(A, r)
gra <- inv_graph(IS = sch)
tidy_inv_graph(gra)

}
\author{
William K. Petry
}
