% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIG.R
\name{ggIG}
\alias{ggIG}
\title{Plot invasion graph using grammar of graphics}
\usage{
ggIG(
  tbl,
  node_size = 18,
  node_stroke = 1,
  node_color = c("grey", "black"),
  node_fill = c(` TRUE` = "#BB5566", `invadable -i TRUE` = "#DDAA33", `saturated -i TRUE`
    = "#004488", ` FALSE` = "#FF9CAA", `invadable -i FALSE` = "#FFD180",
    `saturated -i FALSE` = "#809CDC"),
  edge_color = c("grey80", "grey40")
)
}
\arguments{
\item{tbl}{list containing an invasion graph and associated metadata}

\item{node_size}{numeric, size of the graph nodes/vertices}

\item{node_stroke}{numeric, size of the graph nodes/vertices outline}

\item{node_color}{colors of the graph nodes/vertices outline}

\item{node_fill}{colors of the the graph nodes/vertices fill}

\item{edge_color}{colors of the graph edge}
}
\value{
a ggplot
}
\description{
Plot invasion graph using grammar of graphics
}
\examples{
\dontrun{
n <- 6  # set number of species
A <- -diag(n) - 1.5 * matrix(runif(n^2), n, n)
r <- matrix(1, n, 1)
sch <- inv_scheme(A, r)
gra <- inv_graph(IS = sch)
tidy_gra <- tidy_inv_graph(gra)
ggIG(tidy_gra)
}

}
\author{
William K. Petry
}
