% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggIG.R
\name{ggIG}
\alias{ggIG}
\title{Plot invasion graph using grammar of graphics}
\usage{
ggIG(tbl)
}
\arguments{
\item{tbl}{list containing an invasion graph and associated metadata}
}
\value{
a ggplot
}
\description{
Plot invasion graph using grammar of graphics
}
\examples{
n <- 6  # set number of species
A <- -diag(n) - 1.5 * matrix(runif(n^2), n, n)
r <- matrix(1, n, 1)
sch <- inv_scheme(A, r)
gra <- inv_graph(IS = sch)
ggIG(gra)

}
