% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_scheme.R
\name{inv_scheme}
\alias{inv_scheme}
\title{Compute the invasion scheme for a group of species}
\usage{
inv_scheme(A, r, model = "LV", tol = 1e-14)
}
\arguments{
\item{A}{matrix containing the n x n interaction coefficients between all pairs of
species.}

\item{r}{numeric vector containing the intrinsic growth rates for all n species.}

\item{model}{the underlying dynamic model describing changes in species abundance.
Currently limited to "LV" for Lotka-Volterra.}

\item{tol}{numeric value below which values will be rounded down to zero.}
}
\value{
An n x n numeric matrix containing an invasion scheme, where rows correspond to
    communities and columns correspond to species.
}
\description{
Compute the invasion scheme for a group of species
}
\examples{
n <- 6  # set number of species
A <- -diag(n) - 1.5 * matrix(runif(n^2), n, n)
r <- matrix(1, n, 1)
sch <- inv_scheme(A, r)

}
