% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_scheme.R
\name{inv_scheme}
\alias{inv_scheme}
\title{Compute the invasion scheme for a group of species}
\usage{
inv_scheme(
  A,
  r,
  model = "LV",
  zero = 1e-14,
  tol_singular = 1e-08,
  tol_solve = .Machine$double.eps
)
}
\arguments{
\item{A}{matrix containing the n x n interaction coefficients between all pairs of
species.}

\item{r}{numeric vector containing the intrinsic growth rates for all n species.}

\item{model}{the underlying dynamic model describing changes in species abundance.
Currently limited to "LV" for Lotka-Volterra.}

\item{zero}{numeric absolute value below which invasion scheme growth rates will be
rounded down to zero. Setting equal to 0 will skip rounding. Defaults to 1e-14.}

\item{tol_singular}{numeric tolerance for the test for matrix singularity. Defaults
to 1e-08.}

\item{tol_solve}{numeric tolerance for detecting linear dependencies in the columns of A.
The default is `.Machine$double.eps`.}
}
\value{
An n x n numeric matrix containing an invasion scheme, where rows correspond to
    communities and columns correspond to species.
}
\description{
Compute the invasion scheme for a group of species
}
\examples{
n <- 6  # set number of species
A <- -diag(n) - 1.5 * matrix(runif(n^2), n, n)
r <- matrix(1, n, 1)
sch <- inv_scheme(A, r)

}
\author{
Sebastian Schreiber
}
